imdb_app/requirements.txt
lxml
streamlit
pandas
numpy
plotly
psycopg2-binary
requests
scikit-learn
selenium
webdriver-manager
beautifulsoup4
transformers
torch
flask
